refactor!: prefer direct in-memory db for testing over mocks
since we're using sqlite3, using an in-memory database for testing is
fairly trivial. It helps that the database schema is dead simple.

It also lets me directly use my implementation of the database for
command integration testing.

This makes me a lot more confident in my tests, thus the reason for the
change