add(cmd)!: rewrote `get` command
Testing locally before would affect rummage's go.mod file

To ease the burden of testing locally, a testutil function called
`GoModTidy()` was written to be called at almost every point in `get`'s
tests.

so far i've had no issues with it but if it starts being a problem later
I will address it. so far it's proven to be a small but very effective
solution