3.0.0 (#18)
* prepare for rewrite, make db API more consistent

As of this commit, i'm comitted to rewriting the project for improved
architecture and maintainability.

this commit serves as a foundation for the rewrite, making the database
API a little more consistent, having ALL methods return errors and
rewriting more practical/robust tests

* check err for root command help

* add(db)!: interface for mock testing

- `AddMultiItems()` also now keeps track of items actually added to the
database and returns that amount

* add(cmd): rewrote `populate` command

this is the first time im using mocks for testing so if it doesn't make
sense/is dogshit please forgive me lmao

* add(deps): testify/mock

* ci: adhere to lint rules

* refactor!: prefer direct in-memory db for testing over mocks

since we're using sqlite3, using an in-memory database for testing is
fairly trivial. It helps that the database schema is dead simple.

It also lets me directly use my implementation of the database for
command integration testing.

This makes me a lot more confident in my tests, thus the reason for the
change

* add(cmd)!: rewrote `query` cmd, change `db.UpdateItem()` signature

see commit 416c225a to see changes to the `UpdateItem()` method for the
db.

* add(cmd)!: rewrote/cleaned up `add` cmd

* add(cmd)!: rewrote `remove` cmd

* add(cmd)!: rewrote `get` command

Testing locally before would affect rummage's go.mod file

To ease the burden of testing locally, a testutil function called
`GoModTidy()` was written to be called at almost every point in `get`'s
tests.

so far i've had no issues with it but if it starts being a problem later
I will address it. so far it's proven to be a small but very effective
solution

* fix(ci): fix discrepency issue for `get` test assertions. see comment

To fix the issue where github workflow has a discrepency where
LastAccessed field is asserted to be time.Now().Unix() fails because the
value is off by 1 second, we only can only check the first 8 digits of
the Unix Epoch intead of all 10. These tests do not need such
preciseness

* chore: cleaning up some thing, typos n such

* refactor: making the makefile pretty print!!!

* Update README.md